moviesite/requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2025.7.14
charset-normalizer==3.4.2
dj-rest-auth==7.0.1
django==4.2.23
django-allauth==65.10.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
PyJWT==2.9.0
requests==2.32.4
sqlparse==0.5.3
typing-extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
